renamed the package to standard name
